code/services/process_events/service_code/requirements.txt
# requirements.txt
mysql-connector-python==9.1.0
boto3==1.35.36
boto3-stubs[dynamodb]==1.35.44 # apply this with mypy-boto3-dynamodb for typings
mypy-boto3-dynamodb==1.35.24
boto3-stubs[ecs]==1.35.44
mypy-boto3-secretsmanager==1.35.0
